Replace main script file
